modified Seminar2, Test

+++ src/Seminar1/Homework/Calculator.java
@@ -46,7 +46,7 @@ public class Calculator {
     // Homework
     // Нужно написать в калькуляторе метод вычисления суммы покупки со скидкой и проверить его, используя AssertJ
     // Примерная сигнатура и тело метода:
-    public static double calculatingDiscount(double purchaseAmount, int discountAmount) { // purchaseAmount - сумма покупки, discountAmount - размер скидки
+    public static double calculatingDiscount(double purchaseAmount, double discountAmount) { // purchaseAmount - сумма покупки, discountAmount - размер скидки
         double discountedAmount;
         assertThat(purchaseAmount, discountAmount)
                 .isNotEmpty()
